requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.4.6
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==21.3
httpx==0.27.0
apscheduler==3.10.4
pydantic==2.7.1